Refactor CovidTestList
