+++ appConfig.ts
@@ -7,4 +7,5 @@ export default {
   assessmentVersion: '1.5.0',
   patientVersion: '1.5.0',
   lifestyleVersion: '1.0.1',
+  covidTestVersion: '1.0.0',
 };
