+++ src/features/assessment/CovidTestListScreen.tsx
@@ -1,12 +1,8 @@
 import { RouteProp } from '@react-navigation/native';
 import { StackNavigationProp } from '@react-navigation/stack';
-import moment from 'moment';
 import { Button, Text } from 'native-base';
 import React, { Component } from 'react';
-import { Image, StyleSheet, TouchableOpacity, View } from 'react-native';
-import key from 'weak-key';
-
-import { chevronRight, pending, tick } from '@assets';
+import { StyleSheet, View } from 'react-native';
 import { colors } from '@theme';
 import ProgressStatus from '@covid/components/ProgressStatus';
 import Screen, { Header, ProgressBlock } from '@covid/components/Screen';
@@ -15,8 +11,9 @@ import { AssessmentInfosRequest } from '@covid/core/assessment/dto/AssessmentInf
 import CovidTestService from '@covid/core/user/CovidTestService';
 import { CovidTest } from '@covid/core/user/dto/CovidTestContracts';
 import AssessmentCoordinator from '@covid/core/assessment/AssessmentCoordinator';
-import i18n, { getDayName, getMonthName } from '@covid/locale/i18n';
+import i18n from '@covid/locale/i18n';
 import { assessmentService } from '@covid/Services';
+import { CovidTestRow } from '@covid/components/CovidTestRow/CovidTestRow';
 
 import { ScreenParamList } from '../ScreenParamList';
 
@@ -74,38 +71,6 @@ export default class CovidTestListScreen extends Component<Props, State> {
   render() {
     const currentPatient = AssessmentCoordinator.assessmentData.currentPatient;
 
-    const resultString = (result: string) => {
-      switch (result) {
-        case 'positive':
-          return i18n.t('covid-test-list.positive');
-        case 'negative':
-          return i18n.t('covid-test-list.negative');
-        case 'failed':
-          return i18n.t('covid-test-list.failed');
-        default:
-          return i18n.t('covid-test-list.pending');
-      }
-    };
-
-    const icon = (result: string) => {
-      if (result === 'waiting') {
-        return pending;
-      } else {
-        return tick;
-      }
-    };
-
-    const dateString = (test: CovidTest) => {
-      if (test.date_taken_specific) {
-        const date = moment(test.date_taken_specific).toDate();
-        return `${getMonthName(date)} ${date.getDate()} (${getDayName(date)})`;
-      } else {
-        const stateDate = moment(test.date_taken_between_start).toDate();
-        const endDate = moment(test.date_taken_between_end).toDate();
-        return `${getMonthName(stateDate)} ${stateDate.getDate()} - ${getMonthName(endDate)} ${endDate.getDate()}`;
-      }
-    };
-
     return (
       <View style={styles.rootContainer}>
         <Screen profile={currentPatient.profile} navigation={this.props.navigation}>
@@ -123,22 +88,7 @@ export default class CovidTestListScreen extends Component<Props, State> {
 
           <View style={styles.list}>
             {this.state.covidTests.map((item: CovidTest) => {
-              return (
-                <TouchableOpacity
-                  key={key(item)}
-                  style={styles.itemTouchable}
-                  onPress={() => AssessmentCoordinator.goToAddEditTest(item)}>
-                  <Image source={icon(item.result)} style={styles.tick} />
-                  <RegularText style={item.result === 'waiting' ? styles.pendingText : []}>
-                    {dateString(item)}
-                  </RegularText>
-                  <View style={{ flex: 1 }} />
-                  <RegularText style={item.result === 'waiting' ? styles.pendingText : []}>
-                    {resultString(item.result)}
-                  </RegularText>
-                  <Image source={chevronRight} style={styles.chevron} />
-                </TouchableOpacity>
-              );
+              return <CovidTestRow item={item} />;
             })}
           </View>
         </Screen>
@@ -183,28 +133,10 @@ const styles = StyleSheet.create({
     fontSize: 16,
     color: colors.primary,
   },
-  itemTouchable: {
-    height: 40,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
   list: {
     margin: 16,
   },
-  tick: {
-    marginEnd: 8,
-    height: 16,
-    width: 16,
-  },
-  chevron: {
-    marginStart: 4,
-    height: 12,
-    width: 12,
-  },
   topText: {
     marginTop: 8,
-  },
-  pendingText: {
-    color: colors.secondary,
   },
 });
